Add setter to noLengthLimit in BarcodeGenerator
Add a setter for the attribute noLengthLimit in the
BarcodeGenerator.
In some situations (For example, generating a Barcode in Gs1128),
you need to no limit the length of the barcode, but in the class
BarcodeGenerator yo can't define the value of this property,
which now exists in the "CodeClass"'s.